Don't pass -fvisibility=hidden to clang-cl on Windows
When using clang-cl on windows -fvisibility=hidden is not
an known argument. Instead it behaves exactly like MSVC in
this case. So let's make sure we take that path.

In CMake clang-cl sets both CMAKE_C_COMPILER_ID=clang and
MSVC get's defined since clang-cl is basically a MSVC
emulator. So guarding like we do in this patch seems logical.