Correctly handle linker flags for static libs
Curl CMake was setting the the EXE flags for static libraries
which made the /manifest:no flag ended up when linking the
static library, which is not a valid flag for lib.exe or llvm-lib.exe
and caused llvm-lib to exit with an error.

The better way to handle this is to make sure that we pass the correct
linker flags to CMAKE_STATIC_LINKER_FLAGS instead.